+++ ui/package.json
@@ -21,7 +21,6 @@
     "luxon": "^1.24.1",
     "mkdirp": "^1.0.4",
     "remark": "^12.0.0",
-    "remark-html": "^11.0.2",
     "remark-parse": "^8.0.3",
     "remark-stringify": "^8.1.1",
     "ts-mdast": "^1.0.0",
@@ -53,7 +52,6 @@
     "esm": "^3.2.25",
     "evergreen-ui": "^4.28.0",
     "html-webpack-plugin": "^4.3.0",
-    "mdast-util-to-string": "^1.1.0",
     "mobx": "^5.15.4",
     "mobx-react-lite": "^2.0.7",
     "mocha": "^8.0.1",
