--- ui/src/journal/document/mdast-util-to-hast.d.ts
@@ -1,3 +1,0 @@
-declare module "mdast-util-to-string" {
-  export default function toString(node: any): string;
-}
